% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_modification.R
\name{add_cenancestor_nexus_content}
\alias{add_cenancestor_nexus_content}
\title{Adds LUCA-MRCA branch to NEXUS content}
\usage{
add_cenancestor_nexus_content(trees_content, ptable, n_digits = 16)
}
\arguments{
\item{trees_content}{NEXUS tree file content read using readLines}

\item{ptable}{.log file with a column with the luca_branch information}

\item{n_digits}{number of decimal digits used to print the branch length}
}
\value{
NEXUS tree file content with the LUCA-MRCA branch length added
}
\description{
Adds LUCA-MRCA branch to NEXUS content
}
\keyword{internal}
