% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_modification.R
\name{add_cenancestor_tree}
\alias{add_cenancestor_tree}
\title{Adds LUCA-MRCA branch to APE tree}
\usage{
add_cenancestor_tree(intree, branch_length = NULL, total_depth = NULL)
}
\arguments{
\item{intree}{ape (phylo) tree}

\item{branch_length}{length of the LUCA-MRCA branch (either this or total_depth required).}

\item{total_depth}{total final tree depth (used to calculate branch_length, required if branch_length is not specified)}
}
\value{
phylo tree
}
\description{
Adds LUCA-MRCA branch to APE tree
}
