% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{AICM}
\alias{AICM}
\title{Method-of-moments estimator of AICM}
\usage{
AICM(loglikelihoods)
}
\arguments{
\item{loglikelihoods}{vector of posterior sample of loglikelihoods}
}
\value{
AICM
}
\description{
Posterior simulation-based analogue of Akaike's information criterion (AIC)
through Markov chain Monte Carlo Raftery et al. 2007
}
\details{
Translated from BEAST's logMarginalLikelihoodAICM function
}
