% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{an_support_above}
\alias{an_support_above}
\title{Probability interval of a PDF over a given density (analytical)}
\usage{
an_support_above(dfun, qfun, d, support = c(-Inf, Inf), tol = 1e-08)
}
\arguments{
\item{dfun}{density function}

\item{qfun}{quantile function}

\item{d}{threshold density}

\item{support}{support}

\item{tol}{tolerance from asymptotes}
}
\value{
A 2-element array with the lower and upper values of the interval.
NAs if the PDF volume over d is 0.
}
\description{
Assumes the PDF is unimodal and continuous
}
\seealso{
\code{\link[=an_volume_above]{an_volume_above()}} which finds the PDF's volume within this
interval
}
