% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{an_volume_above}
\alias{an_volume_above}
\title{Volume of a PDF above a given density (analytical)}
\usage{
an_volume_above(dfun, qfun, d, support = c(-Inf, Inf), tol = 1e-08, ...)
}
\arguments{
\item{dfun}{density function}

\item{qfun}{quantile function}

\item{d}{threshold density}

\item{support}{support}

\item{tol}{tolerance from asymptotes}

\item{...}{parameters for qfun and dfun}
}
\value{
the probability of the random variable being between the bounds
determined by d
}
\description{
This function returns the probability of a continuous random variable being
between two bounds determined by a given density. It assumes an unimodal
distribution
}
\seealso{
\code{\link[=est_volume_above]{est_volume_above()}} for its numerical counterpart
}
