% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{apply_samplemap_table}
\alias{apply_samplemap_table}
\title{Resample posterior using a sample map}
\usage{
apply_samplemap_table(samples, map)
}
\arguments{
\item{samples}{list of posterior sample tables}

\item{map}{data.table with two columns. \emph{x_i} indicates the posterior
sample index and \emph{s_i} indicates the index within that sample (i.e.,
\code{samples[[x_i]][s_i]}) for each of the final samples (rows)}
}
\value{
posterior sample (array)
}
\description{
Resample posterior using a sample map
}
\keyword{internal}
