% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{calculate_locus_entropy}
\alias{calculate_locus_entropy}
\title{Calculates the entropy of a single locus using the marginalized approach}
\usage{
calculate_locus_entropy(mat)
}
\arguments{
\item{mat}{matrix \code{[sample,peak]}}
}
\value{
the entropy of a single locus
}
\description{
Calculates the entropy of a single locus using the marginalized approach
}
\seealso{
\code{\link[=get_entropy]{get_entropy()}}
}
\keyword{internal}
