% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{calculate_locus_entropy_ml}
\alias{calculate_locus_entropy_ml}
\title{Calculates the entropy of a single locus using the maximum likelihood peak
assignment}
\usage{
calculate_locus_entropy_ml(mat)
}
\arguments{
\item{mat}{matrix \code{[sample,peak]}}
}
\value{
the entropy of a single locus
}
\description{
Calculates the entropy of a single locus using the maximum likelihood peak
assignment
}
\seealso{
\code{\link[=get_entropy]{get_entropy()}}
}
