% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{calculate_model_averaging_posteriors}
\alias{calculate_model_averaging_posteriors}
\title{Calculate posterior probabilities given all log joint probabilities}
\usage{
calculate_model_averaging_posteriors(log_joint_probabilities)
}
\arguments{
\item{log_joint_probabilities}{array of the log joint probabilities (log marginal likelihoods +
log prior) of each of the posterior samples (x)}
}
\value{
probabilities
}
\description{
Calculate posterior probabilities given all log joint probabilities
}
\keyword{internal}
