% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{calculate_model_averaging_posteriors_file}
\alias{calculate_model_averaging_posteriors_file}
\title{Calculate posterior probabilities given all log joint probabilities}
\usage{
calculate_model_averaging_posteriors_file(
  mle_file,
  mle_method = c("PS", "SS", "HME"),
  n_cells_regex = ".*\\\\.([0-9]+)cells\\\\..*",
  basename_regex = "\\\\.[0-9]+cells\\\\..*",
  log_prior_probabilities = NULL
)
}
\arguments{
\item{mle_file}{filename of the tabular output with marginal likelihood
estimations generated by \link{mcmc_qc} and related functions. It can also be a
list of files and all will be parsed}

\item{mle_method}{sorted list of preferred methods to estimate the marginal
likelihood for model selection. Choose between PS (Path Sampling), SS
(Stepping Stone), and HME (Harmonic Mean Estimator)}

\item{n_cells_regex}{R regular expression with a group that captures the S
number from the MLE.csv filename. Again, the default should work (see
file_regex)}

\item{basename_regex}{R regular expression to substitute with "" to generate
the cond name from the MLE.csv filename. Again, the default should work
(see file_regex)}

\item{log_prior_probabilities}{array of log prior probabilities for each of
the models to average across, named using the corresponding value of S
(S_XX).}
}
\value{
data.table with posterior probabilities for columns S_XX, with two
additional columns indicating patient and method
}
\description{
Calculate posterior probabilities given all log joint probabilities
}
