% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{check_param_prior}
\alias{check_param_prior}
\title{Internal prior-setting checking tool for a given parameter and prior config}
\usage{
check_param_prior(param, config = "O", prior_settings = NULL)
}
\arguments{
\item{param}{name of the parameter}

\item{config}{prior settings used, O by default}

\item{prior_settings}{user-specified prior settings, normally set by
modification of the default prior settings using \code{\link[=user_priors]{user_priors()}}}
}
\value{
prior_settings
}
\description{
Internal prior-setting checking tool for a given parameter and prior config
}
\seealso{
\code{\link[=check_priors]{check_priors()}}
}
