% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{check_priors}
\alias{check_priors}
\title{Finds invalid prior settings}
\usage{
check_priors(prior_settings = NULL)
}
\arguments{
\item{prior_settings}{in case you need to check user prior configuration instead of the default (edge use-case)}
}
\value{
list of configs with a list of invalid priors within
}
\description{
Returns the list of configs with a list of invalid priors within
}
