% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{check_write_csv}
\alias{check_write_csv}
\title{Writes a CSV file if flag is not null}
\usage{
check_write_csv(table, flag, outdir, filename = NULL, canceled_warning = NULL)
}
\arguments{
\item{table}{table to write in CSV}

\item{flag}{print if !is.null. Usually the filename or a filename part}

\item{outdir}{output directory. It will be made recursively if it does not exist, without warnings}

\item{filename}{name of the csv file. If null, flag is used}

\item{canceled_warning}{warning to produce if the file is not written}
}
\value{
TRUE if the file was written, FALSE otherwise
}
\description{
Writes a CSV file if flag is not null
}
\keyword{internal}
