% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{compare_e_a}
\alias{compare_e_a}
\title{Compares an empirical and analytical distribution}
\usage{
compare_e_a(
  x,
  dfun,
  qfun,
  prob = 0.95,
  support = c(-Inf, Inf),
  tol = 1e-08,
  ...
)
}
\arguments{
\item{x}{a vector of samples from the distribution of interest}

\item{dfun}{density function}

\item{qfun}{quantile function}

\item{prob}{cumulative probability at which to find the density threshold}

\item{support}{support}

\item{tol}{tolerance from asymptotes}

\item{...}{other params for \code{\link[ks:kde]{ks::kde()}}}
}
\value{
list with the posterior volume, prior volume, density threshold,
volume ratio, update info (log2 transformation of the former), and
Kullback–Leibler (KL) divergence from the analytical distribution to the
empirical (False)
}
\description{
Used in a Bayesian context to compare Prior and Posterior distributions.
}
\seealso{
\link{is_dispersion_comparable_e_a} for a binary take of this metric, and
\link{kl_divergence_prior_to_posterior} for the other metric used here
}
