% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_default_params.R
\docType{data}
\name{default_model_params}
\alias{default_model_params}
\title{Model Settings}
\format{
A named list with values for
errorModel.deltaOffset,errorModel.etaOffset, errorModel.kappaScale, and
alignment.stemCells
}
\usage{
default_model_params
}
\description{
Database with default model parameter values based on empirical data
}
\details{
It is used internally by \code{\link[=qc_csv]{qc_csv()}} and can be obtained and modified
using \code{\link[=get_model_params]{get_model_params()}}. It is automatically loaded on package load
in zzz.R.
}
\keyword{internal}
