% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_default_settings.R
\docType{data}
\name{default_prior_settings}
\alias{default_prior_settings}
\title{Prior Settings}
\format{
A named list of lists. The first level indicates the configuration
settings (e.g., O for original), the second level indicates the parameter
name, and the third , the dfun, qfun, and a list of parameters needed to
run those functions. If the defaults are not valid, they should contain a
valid element set to FALSE.
}
\usage{
default_prior_settings
}
\description{
Database with default prior configurations.
}
\details{
It is used internally by \code{\link[=get_dfun_prior]{get_dfun_prior()}} and \code{\link[=get_qfun_prior]{get_qfun_prior()}}
and can be obtained and modified using \code{\link[=user_priors]{user_priors()}} and checked with
\code{\link[=check_priors]{check_priors()}}. It is automatically loaded on package load in zzz.R, and
is generated during development using
data_generation/make_prior_default_settings.R (only when needed, using a
hash). The source object is default_prior_settings below.
}
\keyword{internal}
