% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{delete_cached_CSV_data}
\alias{delete_cached_CSV_data}
\title{Deletes CSV data cached by \link{load_Bdata_from_CSVs}}
\usage{
delete_cached_CSV_data()
}
\description{
Deletes CSV data cached by \link{load_Bdata_from_CSVs}
}
\seealso{
\code{\link[=get_Bdata_from_CSVs]{get_Bdata_from_CSVs()}} \code{\link[=load_Bdata_from_CSVs]{load_Bdata_from_CSVs()}}
}
