% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{detect_constants}
\alias{detect_constants}
\title{Detects constant parameters}
\usage{
detect_constants(thedata, params)
}
\arguments{
\item{thedata}{data.table with a column per paramter and row per posterior
sample}

\item{params}{list of columns to consider}
}
\value{
vector of names of parameters with sd == 0
}
\description{
Detects constant parameters
}
\details{
Defined as those parameters with sd = 0
}
