% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_origin.R
\name{emd_to_origin}
\alias{emd_to_origin}
\title{Calculates the distance between the origin and the data}
\usage{
emd_to_origin(freqs)
}
\arguments{
\item{freqs}{empirical frequencies}
}
\value{
list with the Earth's Mover Distance (emd) and the relative Earth's
Mover Distance (rEMD)
}
\description{
It uses the Earth's Mover Distance between the assumed starting frequencies
(0.5 each for fully methylated and fully demethylated) and the empirical
frequencies
}
\details{
The relative Earth's Mover Distance is normalized by the maximum
theoretical distance. Importantly, this is an overestimation since, given a
specific phyfum model, the maximum expected distance is that of the
equilibrium frequencies of the substitution process. We may add this in the
near future but we have not for now.
}
