% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{ensure_list_chains}
\alias{ensure_list_chains}
\title{Ensures that chains are in the expected format}
\usage{
ensure_list_chains(chains)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}
}
\value{
List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values
}
\description{
Ensures that chains are in the expected format
}
\keyword{internal}
