% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{ensure_list_traces}
\alias{ensure_list_traces}
\title{Ensures that traces are in the expected format}
\usage{
ensure_list_traces(traces)
}
\arguments{
\item{traces}{MCMC data. List of \code{\link[=load_traces]{load_traces()}} values}
}
\value{
List of \code{\link[=load_traces]{load_traces()}} values
}
\description{
Ensures that traces are in the expected format
}
\keyword{internal}
