% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{est_volume_above}
\alias{est_volume_above}
\title{Estimates the volume above a density threshold in an empirical density function}
\usage{
est_volume_above(kde_obj, d)
}
\arguments{
\item{kde_obj}{\code{\link[ks:kde]{ks::kde()}} object}

\item{d}{density threshold}
}
\value{
the probability of the random variable being between the bounds
determined by d
}
\description{
Estimates the volume above a density threshold in an empirical density function
}
\seealso{
\code{\link[=an_volume_above]{an_volume_above()}} for its analytical counterpart
}
