% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{estimate_params}
\alias{estimate_params}
\title{Parses a phyfum .log file and summarizes a list of parameters}
\usage{
estimate_params(
  log_file,
  burnin = 0.1,
  sumfun = c("mean", "median"),
  params = c("errorModel.deltaOffset", "errorModel.etaOffset", "errorModel.kappaScale",
    "flipflop.mu", "flipflop.gamma", "flipflop.lambda", "alignment.stemCells")
)
}
\arguments{
\item{log_file}{name of the log file to parse}

\item{burnin}{proportion of samples to be removed as burnin}

\item{sumfun}{summary function. One of mean or median.}

\item{params}{list of parameters to parse}
}
\value{
data.table with a a column per param with its point estimate
}
\description{
Parses a phyfum .log file and summarizes a list of parameters
}
