% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{extract_ptable}
\alias{extract_ptable}
\title{Obtains the parameter trace from from multiple RWTY MCMC samples}
\usage{
extract_ptable(chains, burnin_p, remove_burnin = FALSE)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin. Calculated
by trace and flagged or removed}

\item{remove_burnin}{Logical that indicates if the burnin states should be
removed (True) or flagged (False, Default)}
}
\value{
data.table with same columns as chain$ptable + a chain factor
}
\description{
Obtains the parameter trace from from multiple RWTY MCMC samples
}
\keyword{internal}
