% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{fix_nexus}
\alias{fix_nexus}
\title{Fixes nexus files of unfinished phyfum runs}
\usage{
fix_nexus(file, backup = TRUE)
}
\arguments{
\item{file}{.trees files}

\item{backup}{Logical. Whether a .trees_bkp with the original contents should
be created before fixing the file}
}
\value{
TRUE if the file has been fixed, FALSE if it didn't need fixing
}
\description{
Fixes nexus files of unfinished phyfum runs
}
