% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{fix_nexus_content}
\alias{fix_nexus_content}
\title{Fixes read NEXUS content from unfinished PHYFUM runs}
\usage{
fix_nexus_content(nexus_content)
}
\arguments{
\item{nexus_content}{result from readLines(nexus)}
}
\value{
fixed content
}
\description{
Fixes read NEXUS content from unfinished PHYFUM runs
}
\keyword{internal}
