% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{generate_tree_trace_reftree}
\alias{generate_tree_trace_reftree}
\title{Generates a trace of tree distances from a focal tree}
\usage{
generate_tree_trace_reftree(
  tree_list,
  treedist = c("PD", "RF"),
  tree_i = NULL,
  focal_tree = NULL
)
}
\arguments{
\item{tree_list}{list of trees}

\item{treedist}{Tree distance to be used to calculate the distance to the
focal tree ("PD" or "RF")}

\item{tree_i}{pre-specified element of tree_list that should be used as
reference tree. If NULL (default) it is sampled at random from tree_dist}

\item{focal_tree}{focal tree to use instead of sampling it from the chain}
}
\value{
array of tree distances
}
\description{
To calculate ESS using the pseudo ESS approach. Typically using
\code{\link[=tree_convergence_stats]{tree_convergence_stats()}}
}
\seealso{
\code{\link[=generate_tree_traces]{generate_tree_traces()}} \code{\link[=tree_convergence_stats]{tree_convergence_stats()}}
}
\keyword{internal}
