% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{generate_tree_traces}
\alias{generate_tree_traces}
\title{Generates a tree-topoloy trace}
\usage{
generate_tree_traces(
  chains,
  burnin_p = 0,
  treedist = "PD",
  remove_burnin = TRUE,
  ...
)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin. Calculated
by trace and flagged or removed}

\item{treedist}{Tree distance to be used to calculate the distance to the
focal tree ("PD" or "RF")}

\item{remove_burnin}{Logical that indicates if the burnin states should be
removed (True) or flagged (False, Default)}

\item{...}{
  Arguments passed on to \code{\link[=generate_tree_trace_reftree]{generate_tree_trace_reftree}}
  \describe{
    \item{\code{tree_i}}{pre-specified element of tree_list that should be used as
reference tree. If NULL (default) it is sampled at random from tree_dist}
    \item{\code{focal_tree}}{focal tree to use instead of sampling it from the chain}
  }}
}
\value{
data.table (long format) with a row per MCMC sample and columns for
treedistance and chain
}
\description{
To calculate ESS using the pseudo ESS approach. Typically using
\code{\link[=tree_convergence_stats]{tree_convergence_stats()}}.
}
\details{
Tree_i corresponds to the tree number after burnin if burnin_p>0
}
\seealso{
\code{\link[=generate_tree_trace_reftree]{generate_tree_trace_reftree()}} \code{\link[=tree_convergence_stats]{tree_convergence_stats()}}
}
\keyword{internal}
