% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{get_Bdata_from_CSVs}
\alias{get_Bdata_from_CSVs}
\title{Obtains b-values from CSV files (cached or not)}
\usage{
get_Bdata_from_CSVs(
  csv_dir,
  patient_id,
  sample_pairing,
  matrix = T,
  isParallel = T
)
}
\arguments{
\item{csv_dir}{directory with CSV files to parse}

\item{patient_id}{patient ID to obtain the CSV column names to parse from
sample_pairing}

\item{sample_pairing}{data.table with a patient column and a samplename
column.}

\item{matrix}{whether the return should be a data table (column by sample) or an alignment-like matrix (column by fCpG). All phyfumr functions expect matrix=T (Default)}

\item{isParallel}{True by default, assuming this function is executed in
parallel and warning when this is not safe to do.}
}
\value{
b-value dataset in data.table or matrix form
}
\description{
Obtains b-values from CSV files (cached or not)
}
\details{
CSV files may contain data from multiple patients but the CSV
parsing is done by patient. To avoid re-loading the same files multiple
times, \code{\link[=load_Bdata_from_CSVs]{load_Bdata_from_CSVs()}} can be used first to load all files in memory
and then parse them efficiently here. This function can be used in parallel
if the data has been previously cached. This function assumes it is
executed in parallel and stops when it is not safe to do so. If the user is
certain the function is not run in parallel, isParallel should be set to F.
The cached data can be deleted using \code{\link[=delete_cached_CSV_data]{delete_cached_CSV_data()}}
}
\seealso{
\code{\link[=load_Bdata_from_CSVs]{load_Bdata_from_CSVs()}} \link{delete_cached_CSV_data}
}
