% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{get_Bdata_from_XML}
\alias{get_Bdata_from_XML}
\title{Reads b-values from a Phyfum XML file}
\usage{
get_Bdata_from_XML(xml_file, matrix = T)
}
\arguments{
\item{xml_file}{xml file}

\item{matrix}{whether the return should be a data table (column by sample) or an alignment-like matrix (column by fCpG). All phyfumr functions expect matrix=T (Default)}
}
\value{
b-value dataset in data.table or matrix form
}
\description{
Reads b-values from a Phyfum XML file
}
