% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{get_dfun_prior}
\alias{get_dfun_prior}
\title{Returns the density function of the Prior for a given parameter}
\usage{
get_dfun_prior(param, config = "O", prior_settings = NULL)
}
\arguments{
\item{param}{name of the parameter}

\item{config}{prior settings used, O by default}

\item{prior_settings}{user-specified prior settings, normally set by
modification of the default prior settings using \code{\link[=user_priors]{user_priors()}}}
}
\value{
function that returns the density of the parameter value x, or
function that returns NA_real_ if it is unknown
}
\description{
Returns the density function of the Prior for a given parameter
}
\seealso{
\code{\link[=check_priors]{check_priors()}}
}
