% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_origin.R
\name{get_empirical_freqs}
\alias{get_empirical_freqs}
\title{Calculates empirical frequencies given the tip likelihoods}
\usage{
get_empirical_freqs(likelihoods, ml = F)
}
\arguments{
\item{likelihoods}{3D array with \code{[sample,peak,locus]} dimensions with
the b-peak likelihoods for each sample and fCpG locus as returned by
\code{\link[=get_peak_likelihoods]{get_peak_likelihoods()}}}

\item{ml}{Logical that specifies whether the entropy is calculated using
maximum likelihood peak assignments or marginalized over all possible peaks}
}
\value{
empirical frequencies
}
\description{
Calculates empirical frequencies given the tip likelihoods
}
\seealso{
\code{\link[=get_peak_likelihoods]{get_peak_likelihoods()}}
}
