% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Calculates the b-value entropy under they phyfum model}
\usage{
get_entropy(likelihoods, ml = T)
}
\arguments{
\item{likelihoods}{3D array with \code{[sample,peak,locus]} dimensions with
the b-peak likelihoods for each sample and fCpG locus as returned by
\code{\link[=get_peak_likelihoods]{get_peak_likelihoods()}}}

\item{ml}{Logical that specifies whether the entropy is calculated using
maximum likelihood peak assignments or marginalized over all possible peaks}
}
\value{
expected entropy
}
\description{
Calculates the b-value entropy under they phyfum model
}
\details{
The full/marginalized likelihood calculation does not work very well
empirically and thus we recommend using the ML implementation.
}
