% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_fss_entropy}
\alias{get_fss_entropy}
\title{Compute full substitution saturation entropy}
\usage{
get_fss_entropy(freqs, n, delta, eta, kappa, ml = T, ...)
}
\arguments{
\item{freqs}{Vector of category probabilities (i.e., frequencies, sums to 1)}

\item{n}{Total count (number of sequences)}

\item{delta}{flipflop.delta parameter of the phyfum model (methylated
background noise)}

\item{eta}{flipflop.eta parameter of the phyfum model (de-methylated
background negative noise)}

\item{kappa}{flipflop.kappa parameter of the phyfum model (b concentration)}

\item{ml}{Logical that specifies whether the entropy is calculated using
maximum likelihood peak assignments or marginalized over all possible peaks}

\item{...}{
  Arguments passed on to \code{\link[=get_fss_entropy_ml]{get_fss_entropy_ml}}
  \describe{
    \item{\code{prob_threshold}}{Absolute multinomial probability cutoff (optional). Set
to 0 to explore all compositions (Default).}
  }}
}
\value{
list with the expected full saturation ML entropy estimator,
expected full saturation ML entropy estimator without the
skipped-probability normalization, expected full saturation skipped
probability, probability threshold, and max_error
}
\description{
Old version that should not be used currently. Use \code{\link[=get_fss_entropy_ml]{get_fss_entropy_ml()}}
instead.
}
\details{
This is an old version that allows to use ML or full-probability.
The full-probability only uses compositions and simulation and does not
include probability filtering
}
\seealso{
\code{\link[=get_fss_entropy_ml]{get_fss_entropy_ml()}}
}
\keyword{internal,}
\keyword{unsupported}
