% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_fss_entropy_ml_full_composition}
\alias{get_fss_entropy_ml_full_composition}
\title{Compute full substitution saturation entropy}
\usage{
get_fss_entropy_ml_full_composition(freqs, n, prob_threshold = 0)
}
\arguments{
\item{freqs}{Vector of category probabilities (i.e., frequencies, sums to 1)}

\item{n}{Total count (number of sequences)}

\item{prob_threshold}{Absolute multinomial probability cutoff (optional). Set
to 0 to explore all compositions (Default).}
}
\value{
list with the expected full saturation ML entropy estimator,
expected full saturation ML entropy estimator without the
skipped-probability normalization, expected full saturation skipped
probability, probability threshold, and max_error
}
\description{
This version only works well with small n and length(freqs)
}
\details{
This version uses full enumeration of multinomial compositions and
thus requires a lot of memory when freqs or n are large. The recursive
version \code{\link[=get_fss_entropy_ml]{get_fss_entropy_ml()}} is usually preferred. Filtering compositions
with very small probability speeds up the computation without big effect in
the estimation.
}
\seealso{
[\code{\link[=get_fss_entropy_ml]{get_fss_entropy_ml()}}]
}
