% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{get_function}
\alias{get_function}
\title{Obtains function from function name including package/environment name}
\usage{
get_function(x)
}
\arguments{
\item{x}{the name of the function. E.g., stats::qfun}
}
\value{
function
}
\description{
Obtains function from function name including package/environment name
}
