% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{get_kde_threshold}
\alias{get_kde_threshold}
\title{Finds a density such that the cumulative probability of an empirical density
function is above a given value}
\usage{
get_kde_threshold(x, prob = 0.95, ...)
}
\arguments{
\item{x}{a vector of samples from the distribution of interest}

\item{prob}{cumulative probability at which to find the density threshold}

\item{...}{other params for \code{\link[ks:kde]{ks::kde()}}}
}
\value{
a list with the density threshold and the kde object
}
\description{
Finds a density such that the cumulative probability of an empirical density
function is above a given value
}
\details{
KDE-based
}
