% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_model_params}
\alias{get_model_params}
\title{Returns the default model settings for user exploration and modification}
\usage{
get_model_params()
}
\value{
A named list with values for
errorModel.deltaOffset,errorModel.etaOffset, errorModel.kappaScale, and
alignment.stemCells
}
\description{
Returns the default model settings for user exploration and modification
}
\seealso{
\code{\link[=qc_csv]{qc_csv()}}
}
