% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_peak_likelihoods}
\alias{get_peak_likelihoods}
\title{Calculates the b-peak likelihoods given a fCpG dataset and phyfum model}
\usage{
get_peak_likelihoods(b_data, delta, eta, kappa, s)
}
\arguments{
\item{b_data}{b-value alignment (i.e., matrix (or 2-dimensional array) with a
column per fCpG site and row per sample)}

\item{delta}{flipflop.delta parameter of the phyfum model (methylated
background noise)}

\item{eta}{flipflop.eta parameter of the phyfum model (de-methylated
background negative noise)}

\item{kappa}{flipflop.kappa parameter of the phyfum model (b concentration)}

\item{s}{s parameter of the phyfum model (number of stem cells)}
}
\value{
3D array with \code{[sample,peak,locus]} dimmentions with the b-peak
likelihoods for each sample and fCpG locus
}
\description{
Calculates the b-peak likelihoods given a fCpG dataset and phyfum model
}
