% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{get_qfun_prior}
\alias{get_qfun_prior}
\title{Returns the quantile function of the Prior for a given parameter}
\usage{
get_qfun_prior(param, config = "O", prior_settings = NULL)
}
\arguments{
\item{config}{prior settings used, O by default}

\item{prior_settings}{user-specified prior settings, normally set by
modification of the default prior settings using \code{\link[=user_priors]{user_priors()}}}

\item{parameter}{name of the parameter}
}
\value{
function that returns the quantile of the parameter probability p, or
function that returns NA_real_ if it is unknown
}
\description{
Returns the quantile function of the Prior for a given parameter
}
\seealso{
\code{\link[=check_priors]{check_priors()}}
}
