% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{get_summary_tree}
\alias{get_summary_tree}
\title{Uses BEAST's treeannotator to summarize the posterior distribution of trees
from a file}
\usage{
get_summary_tree(
  trees_file,
  burnin_p = 0.1,
  burnin_trees = NULL,
  outname = NULL,
  sum_fun = c("mean", "median", "keep", "ca"),
  treeannotator_args = NULL
)
}
\arguments{
\item{trees_file}{trees files to summarize}

\item{burnin_p}{proportion of samples to be discarded as burnin}

\item{burnin_trees}{number of samples to be discarded as burnin
(burnin_trees has priority)}

\item{outname}{name of the output summary tree file. NULL by default (not
save it)}

\item{sum_fun}{Name of the function to use to summarize node heights}

\item{treeannotator_args}{An array of additional parameters to use when calling
treeannotator.}
}
\value{
the summary tree as an ape::phylo
}
\description{
Uses BEAST's treeannotator to summarize the posterior distribution of trees
from a file
}
\details{
This function uses treeannotator's defaults except for the branch
length summary function. This means that using different versions of
treeannotator may result in different estimates (e.g., former versions
defaulted to MCC trees while newer versions default to hipstr)
}
