% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{get_summary_tree_from_trees}
\alias{get_summary_tree_from_trees}
\title{Uses BEAST's treeannotator to summarize the posterior distribution of trees
from memory}
\usage{
get_summary_tree_from_trees(
  trees,
  temp_name = "temptree",
  burnin_p = 0.1,
  burnin_trees = NULL,
  outname = NULL,
  sum_fun = c("mean", "median", "keep", "ca"),
  treeannotator_args = NULL
)
}
\arguments{
\item{trees}{trees files to summarize}

\item{temp_name}{name of the temporary tree file. If this function is used in
parallel, each thread should use a different temp_name to avoid issues}

\item{burnin_p}{proportion of samples to be discarded as burnin}

\item{burnin_trees}{number of samples to be discarded as burnin
(burnin_trees has priority)}

\item{outname}{name of the output summary tree file. NULL by default (not
save it)}

\item{sum_fun}{Name of the function to use to summarize node heights}

\item{treeannotator_args}{An array of additional parameters to use when calling
treeannotator.}
}
\value{
the summary tree as an ape::phylo
}
\description{
Uses BEAST's treeannotator to summarize the posterior distribution of trees
from memory
}
\details{
\link{get_summary_tree} wrapper that deals with temp files since
treeannotator cannot read from stdin. If this function is used in parallel,
different threads should use different temp_names to avoid collisions.
}
