% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{is_dispersion_comparable_e_a}
\alias{is_dispersion_comparable_e_a}
\title{Compares the densities of two distributions to deem them comparable}
\usage{
is_dispersion_comparable_e_a(
  x,
  dfun,
  qfun,
  prob = 0.95,
  support = c(-Inf, Inf),
  tol = 1e-08,
  ...
)
}
\arguments{
\item{x}{a vector of samples from the distribution of interest}

\item{dfun}{density function}

\item{qfun}{quantile function}

\item{prob}{cumulative probability at which to find the density threshold}

\item{support}{support}

\item{tol}{tolerance from asymptotes}

\item{...}{other params for \code{\link[ks:kde]{ks::kde()}}}
}
\value{
logical indicating if the distributions are comparable (True) or not
(False)
}
\description{
Used in a Bayesian context to compare Prior and Posterior distributions.
Empirically, we found that if the density threshold for a high credible
interval (e.g. 95\%) does not have any volume in the prior (i.e., the prior is
much more disperse than the posterior), the data has informed the posterior
enough and the model is quite robust to changes in the prior.
}
\seealso{
\code{\link[=compare_e_a]{compare_e_a()}} to actually calculate the amount of change in volume
over a density threshold (continuous take of this binary statistic)
}
