% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prior_comparison.R
\name{kl_divergence_prior_to_posterior}
\alias{kl_divergence_prior_to_posterior}
\title{Calculates the Kullback–Leibler (KL) divergence from an empirical distribution to a probability distribution with known density function}
\usage{
kl_divergence_prior_to_posterior(posteriorKDE, dfun)
}
\arguments{
\item{posteriorKDE}{\code{\link[ks:kde]{ks::kde()}} object with the KDE estimation of the posterior
samples}

\item{dfun}{density function that takes a vector of values and returns their
density}
}
\value{
KL divergence in bits
}
\description{
Calculates the Kullback–Leibler (KL) divergence from an empirical distribution to a probability distribution with known density function
}
