% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsing.R
\name{load_Bdata_from_CSV_dirs}
\alias{load_Bdata_from_CSV_dirs}
\title{Caches b-values stored in all CSV files in a list of directories}
\usage{
load_Bdata_from_CSV_dirs(csv_dirs)
}
\arguments{
\item{csv_dirs}{list of directories for which to read all CSV files}
}
\description{
Caches b-values stored in all CSV files in a list of directories
}
\details{
CSV files may contain data from multiple patients but the CSV
parsing is done by patient \code{\link[=get_Bdata_from_CSVs]{get_Bdata_from_CSVs()}}. To avoid re-loading the
same files multiple times, this function can be used first to load all
files in memory and then parse them efficiently. This function must not be
used in parallel since it stores the data in a package environment. All
downstream analyses may be conducted in parallel.
}
\seealso{
\code{\link[=load_Bdata_from_CSVs]{load_Bdata_from_CSVs()}}
}
