% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{load_condition_traces}
\alias{load_condition_traces}
\title{Load PHYFUM posterior samples from the same condition}
\usage{
load_condition_traces(trees_files, log_files = NULL, burnin_p = 0.1)
}
\arguments{
\item{trees_files}{array of PHYFUM's output .trees files (one per independent
run)}

\item{log_files}{array of PHYFUM's output .log files (one per independent
run). If not used, this function will assume they have the same name and
location as their .trees counterparts}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}
}
\value{
list of traces (list of trees and ptable)
}
\description{
If more than one run was generated per condition, they will be merged
}
\keyword{internal}
