% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{load_trace}
\alias{load_trace}
\title{Loads PHYFUM MCMC traces}
\usage{
load_trace(treefile, logfile = NULL, burnin_p = 0.1)
}
\arguments{
\item{treefile}{A path to a tree file containing an MCMC chain of trees}

\item{logfile}{A path to a file containing model parameters and likelihoods.
If no path is provided, it looks for a file with the same name as the tree
file with .log extension}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}
}
\value{
output A phyfumr_trace object containing the multiPhylo and the table
of values from the log file if available.
}
\description{
Loads trees, looks for a log file of tree likelihoods and parameter values,
returns an phyfumr_trace object containing both
}
\details{
Automatically fixes NEXUS files of unfinished runs on the fly
}
\seealso{
\code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}},
\code{\link[rwty]{load.trees}}
}
