% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{load_traces}
\alias{load_traces}
\title{Load phyfum posterior samples}
\usage{
load_traces(trees_files, log_files = NULL, burnin_p = 0, ...)
}
\arguments{
\item{trees_files}{list of phyfum's output .trees files (one per independent
run)}

\item{log_files}{list of phyfum's output .log files (one per independent
run). If not used, this function will assume they have the same name and
location as their .trees counterparts}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}

\item{...}{
  Arguments passed on to \code{\link[=fix_nexus]{fix_nexus}}
  \describe{
    \item{\code{backup}}{Logical. Whether a .trees_bkp with the original contents should
be created before fixing the file}
  }}
}
\value{
list of RWTY's chains
}
\description{
Load phyfum posterior samples
}
\keyword{internal}
