% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{mcmc_qc_condition}
\alias{mcmc_qc_condition}
\title{Complete MCMC QC per condition (patient and S parameter value)}
\usage{
mcmc_qc_condition(
  trees_files,
  log_files = NULL,
  out_dir,
  plot_dir,
  burnin_p = 0.1,
  min_ess = 200,
  max_rhat = 1.1,
  cred_mass = 0.95,
  n_focal_trees = 20,
  treedist = c("PD", "RF"),
  n_cores = NULL,
  base_name = NULL,
  problematic_output_suffix = "problematicParams.csv",
  all_output_suffix = "allParams.csv",
  posterior_plots_suffix = "plotPosteriorDensities.pdf",
  rwty_plots_suffix = "plotsRWTY.pdf",
  correlation_plots_suffix = "plotCorrelations.jpeg",
  mle_suffix = "MLE.csv",
  ...
)
}
\arguments{
\item{trees_files}{list of phyfum's output .trees files (one per independent
run)}

\item{log_files}{list of phyfum's output .log files (one per independent
run). If not used, this function will assume they have the same name and
location as their .trees counterparts}

\item{out_dir}{directory for tabular outputs}

\item{plot_dir}{directory for graphical outputs}

\item{burnin_p}{proportion of input traces to be discarded as burnin}

\item{min_ess}{minimum ESS. Parameters with ESS under this value will be
flagged as problematic}

\item{max_rhat}{maximum Rhat \code{\link[rstan:Rhat]{rstan::Rhat()}}. Parameters with Rhat higher
than this will be flagged as problematic}

\item{cred_mass}{credibility mass to determine high density intervals}

\item{n_focal_trees}{number of focal trees to calculate a tree distance to
estimate topological convergence using the pseudo-ESS method}

\item{treedist}{tree distance used for the topological pseudo-ESS}

\item{n_cores}{if NULL, using 1 core for rwty and let's data.table autodetect
the number of cores (Default, good to run this function in series).
Otherwise, it fixes the number of cores that rwty and data.table will use.
They never run in parallel so both can use the maximum number of cores}

\item{base_name}{common name to use as prefix for all outputs to share output
folders between patients. By default, the name of the .trees files is used}

\item{problematic_output_suffix}{suffix for the filename of the tabular
output with a list of parameters that may have mixing/convergence problems.
If NULL, this output is not generated.}

\item{all_output_suffix}{suffix for the filename of the tabular output with
all mixing and convergence statistics. If NULL, this output is not
generated.}

\item{posterior_plots_suffix}{suffix for the marginal density plots file. If
NULL, this output is not generated.}

\item{rwty_plots_suffix}{suffix for the RWTY plots. If NULL, this output is
not generated.}

\item{correlation_plots_suffix}{suffix for the RWTY correlation plots. If
NULL, this output is not generated.}

\item{mle_suffix}{suffix for the tabular output with marginal likelihood
estimations using HME and AICm methods.}

\item{...}{
  Arguments passed on to \code{\link[=fix_nexus]{fix_nexus}}
  \describe{
    \item{\code{backup}}{Logical. Whether a .trees_bkp with the original contents should
be created before fixing the file}
  }}
}
\value{
list with data.tables with all = convergence statistics, problematic = parameter with problems, MLE = mle estimation table
}
\description{
This function generates plots and tables to aid assessing phyfum's mixing and
convergence. It can use several chains per run, and automatically flags some
issues using Rhat and ESS.
}
\details{
When n_cores is > 1, RNGkind("L'Ecuyer-CMRG") random number
generator is used instead of the default. This allows mclapply to be
reproducible.
}
