% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{mcmc_qc_gather}
\alias{mcmc_qc_gather}
\title{Gather and summarize MCMC QC per experiment}
\usage{
mcmc_qc_gather(
  out_dir,
  problematic_output_suffix = "problematicParams.csv",
  all_output_suffix = "allParams.csv",
  mle_suffix = "MLE.csv",
  problematic_output = "problematicParams.csv",
  all_output = "allParams.csv",
  mle_output = "MLE.csv",
  conditions_with_issues_output = "problematicConditions.csv"
)
}
\arguments{
\item{out_dir}{output directory where all the \code{\link[=mcmc_qc_condition]{mcmc_qc_condition()}} outputs
are located, which will also be where the output of this function will be
located (as \code{\link[=mcmc_qc]{mcmc_qc()}})}

\item{problematic_output_suffix}{suffix for the filename of the tabular
output with a list of parameters that may have mixing/convergence problems.
If NULL, this output is not generated.}

\item{all_output_suffix}{suffix for the filename of the tabular output with
all mixing and convergence statistics. If NULL, this output is not
generated.}

\item{mle_suffix}{suffix for the tabular output with marginal likelihood
estimations using HME and AICm methods.}

\item{problematic_output}{filename of the tabular output with a list of
parameters that may have mixing/convergence problems. If NULL, this output
is not generated.}

\item{all_output}{filename of the tabular output with all mixing and
convergence statistics. If NULL, this output is not generated..}

\item{mle_output}{filename of the tabular output with marginal likelihood
estimations using HME and AICm methods.}

\item{conditions_with_issues_output}{output filename for a summary table that
lists conditions with topological or continuous parameter convergence
issues. If NULL, this output is not generated.}
}
\value{
list with data.tables with all = convergence statistics, problematic
= parameter with problems, MLE = mle estimation table data.table. Each has
a column indicating the condition (i.e., filename of the input file without
.trees)
}
\description{
This function collects and summarizes the results from \code{\link[=mcmc_qc_condition]{mcmc_qc_condition()}}.
It is an alternative to \code{\link[=mcmc_qc]{mcmc_qc()}} when \code{\link[=mcmc_qc_condition]{mcmc_qc_condition()}} is run in
parallel using multiple computers (typicaly an HPC cluster).
}
\details{
all .trees files within the objective directory will be analized.
Files with the same name will be considered independent phyfum runs under
the same conditions.
}
