% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{merge_traces}
\alias{merge_traces}
\title{Merges a list of RWTY chains into one}
\usage{
merge_traces(chains, burnin_p = 0)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin. Calculated
by trace and flagged or removed}
}
\value{
rwty.chain with the chains concatenated and the burnin states (trees
and parameters) removed.
}
\description{
Merges a list of RWTY chains into one
}
\details{
It assumes samples are taken at regular intervals to perform the
burnin
}
\keyword{internal}
