% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{model_average}
\alias{model_average}
\title{Marginalize posteriors over S across the whole experiment}
\usage{
model_average(
  file_dir,
  out_dir,
  out_tree_dir,
  out_trace_dir,
  mle_file = NULL,
  mle_method = c("PS", "SS", "HME"),
  log_prior_probabilities = NULL,
  n_marginalized_posterior_samples = 10000,
  max_expected_samples = 1,
  warn_biased_marginalization = T,
  burnin_p = 0.1,
  cred_mass = 0.95,
  summarize_trees = TRUE,
  tree_branch_sum_fun = c("mean", "median", "keep", "ca"),
  n_cells_regex = ".*\\\\.([0-9]+)cells\\\\..*",
  basename_regex = "\\\\.[0-9]+cells\\\\..*",
  summary_tree_suffix = ".tree",
  summary_params_suffix = ".csv",
  n_cores = 1,
  precision = NULL,
  ...
)
}
\arguments{
\item{file_dir}{directory that contains .trees and .log files (with or
without subdirectories).}

\item{out_dir}{directory that will contain a summary statistics table per
patient. NULL to disable summary statistics.}

\item{out_tree_dir}{directory that will contain a summary tree per patient.
NULL to disable writing the summary tree (it will still be calculated
unless summarize_trees = FALSE).}

\item{out_trace_dir}{directory that will contain the resampled trace (.trees
and .log files) marginalized over S. NULL to disable trace storage.}

\item{mle_file}{filename of the tabular output with marginal likelihood
estimations generated by \link{mcmc_qc} and related functions. It can also be a
list of files and all will be parsed}

\item{mle_method}{sorted list of preferred methods to use the estimated
marginal likelihood for model selection. Choose between PS (Path Sampling),
SS (Stepping Stone), and HME (Harmonic Mean Estimator). Only useful when
mle_file is provided.}

\item{log_prior_probabilities}{array of log prior probabilities for each of
the models to average across, named using the corresponding value of S
(S_XX).}

\item{n_marginalized_posterior_samples}{length of the resampled posterior. If
NULL, it is maximized sampling without resampling. If more samples than
existing are required, sampling is done with replacement, otherwise
without.}

\item{max_expected_samples}{maximum allowed expected number of samples in an
extreme value of S}

\item{warn_biased_marginalization}{Logical, indicating whether to warn or not
if the limit of
max_expected_samples in an extreme value of S is violated}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}

\item{cred_mass}{credibility mass to determine high density intervals}

\item{summarize_trees}{boolean indicating if trees must be summarized (e.g.,
calculating the MCC tree) or not.}

\item{tree_branch_sum_fun}{summarization function for summary tree branches
(choose from mean, median, keep, and ca). See \link{get_summary_tree}}

\item{n_cells_regex}{R regular expression with a group that captures the S
number from the MLE.csv filename. Again, the default should work (see
file_regex)}

\item{basename_regex}{R regular expression to substitute with "" to generate
the cond name from the MLE.csv filename. Again, the default should work
(see file_regex)}

\item{summary_tree_suffix}{suffix to the patient ID to be used as filename
when storing a summary tree}

\item{summary_params_suffix}{suffix to the patient ID to be used as filename
when storing a summary of continuous parameters}

\item{n_cores}{number of cores to use for coarse-grained parallelism (per S
value per patient)}

\item{precision}{maximum difference between two floating point numbers to be
considered equal (used to check posterior probabilities)}

\item{...}{
  Arguments passed on to \code{\link[=get_summary_tree_from_trees]{get_summary_tree_from_trees}}, \code{\link[=get_summary_tree]{get_summary_tree}}
  \describe{
    \item{\code{treeannotator_args}}{An array of additional parameters to use when calling
treeannotator.}
  }}
}
\value{
List with data.tables with all = convergence statistics, problematic
= parameter with problems, MLE = mle estimation table data.table. Each has
a column indicating the condition (i.e., filename of the input file without
.trees)
}
\description{
This function uses Bayesian Model Averaging to marginalize over S. It can
generate posterior samples, continuous parameter estimates, and point tree
estimates.
}
\details{
All .trees files within the objective directory (and
sub-directories) will be analyzed. Files with the same name will be
considered independent phyfum runs under the same conditions and will be
merged after burning before re-sampling.
}
