% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{model_average_gather}
\alias{model_average_gather}
\title{Gather an summarize the marginalization of posteriors over S}
\usage{
model_average_gather(
  out_dir,
  n_cells_regex = ".*\\\\.([0-9]+)cells\\\\..*",
  basename_regex = "\\\\.[0-9]+cells\\\\..*",
  summary_tree_suffix = "summary.tree",
  summary_params_suffix = "summary.csv",
  pp_suffix = "pp.csv",
  patient_id_separator = "_",
  all_summary_params_filename = "summary.csv",
  pp_filename = "pp.csv"
)
}
\arguments{
\item{out_dir}{directory that will contain a summary statistics table per
patient, and a posterior probability summary. NULL to disable these
outputs.}

\item{n_cells_regex}{R regular expression with a group that captures the S
number from the MLE.csv filename. Again, the default should work (see
file_regex)}

\item{basename_regex}{R regular expression to substitute with "" to generate
the cond name from the MLE.csv filename. Again, the default should work
(see file_regex)}

\item{summary_tree_suffix}{suffix to the patient ID to be used as filename
when storing a summary tree}

\item{summary_params_suffix}{suffix to the patient ID to be used as filename
when storing a summary of continuous parameters}

\item{pp_suffix}{suffix to the patient ID to be used as filename when storing
the posterior probabilities of S. Not written if NULL.}

\item{patient_id_separator}{separator to merge the patient ID with the
suffixes above}

\item{all_summary_params_filename}{filename for the summary of all continuous
parameters across patients (within out_dir)}

\item{pp_filename}{filename for the summary of the posterior probabilities of
S (within out_dir)}
}
\value{
List with ptable: data.table with continuous parameter summary,
pptable: data.table with posterior probabilities, trees: multiPhylo with
all the summary trees.
}
\description{
Run after \link{model_average_patient} for all patients, generates the same
experiment-wide outputs as \link{model_average}.
}
