% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{names_from_objects}
\alias{names_from_objects}
\title{Gets the variable names from a list}
\usage{
names_from_objects(...)
}
\arguments{
\item{...}{objects to get the name of}
}
\value{
vector of names
}
\description{
Gets the variable names from a list
}
