% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{param_convergence_stats}
\alias{param_convergence_stats}
\title{Generates a table with convergence statistics for all continuous parameters}
\usage{
param_convergence_stats(
  thedata,
  params = NULL,
  cred_mass = 0.95,
  by_chain = TRUE,
  cl = 1
)
}
\arguments{
\item{thedata}{MCMC posterior samples in data.table format after
\code{\link[=extract_ptable]{extract_ptable()}}}

\item{params}{List of parameter names to analyze. NULL if all are wanted
(default). In that case, known phyfum output columns that are not
parameters (chain, state), and constant parameters as in
\code{\link[=detect_constants]{detect_constants()}} are removed}

\item{cred_mass}{mass of the selected high posterior density interval (HDI)}

\item{by_chain}{Calculates Rhat, and the two ESS parameters both using all
chains and by chain (otherwise, using all of them only)}

\item{cl}{cl settings for \code{\link[pbapply:pbapply]{pbapply::pblapply()}} if available}
}
\value{
data.table with a row per parameter and columns with convergence
statistics from \code{\link[=convergence_stats]{convergence_stats()}}
}
\description{
Generates a table with convergence statistics for all continuous parameters
}
\seealso{
\code{\link[=extract_ptable]{extract_ptable()}}
}
