% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyfumr-package.R
\docType{package}
\name{phyfumr-package}
\alias{phyfumr}
\alias{phyfumr-package}
\title{phyfumr: R Functions for Analysis and QC of PhyFum Results}
\description{
PHYFUM and PHYFUMflow perform phylogenetic estimation of somatic samples from tissues organized in clonal subpopulations using fluctuating CpG (fCpG) sites of methylation arrays. This package contains functions to QC selected fCpG sites before phylogenetic estimation and QC the estimation results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamallo/phyfumr}
  \item Report bugs at \url{https://github.com/adamallo/phyfumr/issues}
}

}
\author{
\strong{Maintainer}: Diego Mallo \email{dmalload@asu.edu} (\href{https://orcid.org/0000-0002-9046-8859}{ORCID})

}
\keyword{internal}
