% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{plot_continuous_parameters}
\alias{plot_continuous_parameters}
\title{Plots posterior sample and estimated values for a parameter}
\usage{
plot_continuous_parameters(thedata, params, ndigits = 3)
}
\arguments{
\item{thedata}{data.table, with burnin removed if needed}

\item{params}{array of parameter names}

\item{ndigits}{number of digits to round numbers printed in the plot}
}
\value{
list of ggplot2 plots
}
\description{
Plots posterior sample and estimated values for a parameter
}
\keyword{internal}
