% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCpG_qc.R
\name{qc_csv}
\alias{qc_csv}
\title{QC of selected fCpG sites}
\usage{
qc_csv(
  csvs,
  patientData,
  max_rel_emd_to_origin = 0.5,
  max_rel_entropy = 0.95,
  model_params = NULL,
  min_multinomial_p = 0.95,
  ml_entropies = T,
  cl = 1,
  ...
)
}
\arguments{
\item{csvs}{csv file(s) containing b-values. A row per CpG site and column
per sample, with row and column names.}

\item{patientData}{data.table with columns for samplename and patient. The
samplename columns should correspond with the column/sample name in the CSV
or XML files}

\item{max_rel_emd_to_origin}{maximum mean relative earth's mover distance to
the origin}

\item{max_rel_entropy}{maximum mean relative entropy}

\item{model_params}{user_defined model parameters used for the computation.
Obtain the defaults using \code{\link[=get_model_params]{get_model_params()}} and modify them to your
needs.}

\item{min_multinomial_p}{minimum multinomial probability to be included in
the calculation of the full saturation entropy}

\item{ml_entropies}{whether the entropy calculation should be carried out
using maximum likelihood state asignment or not. See \code{\link[=get_entropy]{get_entropy()}}}

\item{cl}{cl settings for \code{\link[pbapply:pbapply]{pbapply::pblapply()}} if available}

\item{...}{arguments for estimate_prob_threshold}
}
\value{
data.table with distance to the origin and relative entropy data per
patient, including if they are suspicious of being problematic (outside the
max* thresholds)
}
\description{
QC of selected fCpG sites
}
\details{
This function can runs in parallel (per patient) if pbapply is
available
}
