% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_default_settings.R
\name{qhalft}
\alias{qhalft}
\title{Quantile function for the half Student's T distribution}
\usage{
qhalft(p, nu, scale)
}
\arguments{
\item{p}{as in \link[LaplacesDemon:dist.Halft]{LaplacesDemon::qhalft}}

\item{nu}{as in \link[LaplacesDemon:dist.Halft]{LaplacesDemon::qhalft}}

\item{scale}{as in \link[LaplacesDemon:dist.Halft]{LaplacesDemon::qhalft}}
}
\value{
the quantile of the half Student's T with p = p
}
\description{
\link{LaplacesDemon} qhalft implementation does not work well for our purposes
here (there's some stochasticity in the results and their variance is quite
large)
}
