% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{remove_burnin}
\alias{remove_burnin}
\title{Removes burnin states in RWTY chain lists}
\usage{
remove_burnin(chains, burnin_p)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}
}
\value{
list of RWTY chains with burnin samples removed
}
\description{
Removes burnin states in RWTY chain lists
}
\keyword{internal}
