% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traces.R
\name{remove_burnin_traces}
\alias{remove_burnin_traces}
\title{Removes burnin states in lists of posterior samples}
\usage{
remove_burnin_traces(traces, burnin_p)
}
\arguments{
\item{traces}{MCMC data. List of \code{\link[=load_traces]{load_traces()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}
}
\value{
list of PHYFUMr traces with burnin samples removed
}
\description{
Removes burnin states in lists of posterior samples
}
\keyword{internal}
