% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{resample_posterior}
\alias{resample_posterior}
\title{Resample posteriors according to probabilities}
\usage{
resample_posterior(x, probs, n_marginalized_posterior_samples = NULL)
}
\arguments{
\item{x}{list of posterior sample tables. They do not have to have the same
length}

\item{probs}{array of sampling probabilities (one per posterior sample x)}

\item{n_marginalized_posterior_samples}{length of the resampled posterior. If
NULL, it is maximized sampling without resampling. If more samples than
existing are required, sampling is done with replacement, otherwise
without.}
}
\value{
a list with the posterior sample and a 2D array with a row per final
posterior sample and column per input posterior sample, with elements
indicating the ith position of the input posterior sample to be used as
output, or 0 otherwise. Only one element >0 per row is expected
}
\description{
Resample posteriors according to probabilities
}
\keyword{internal}
