% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{run}
\alias{run}
\title{Run a bundled script}
\usage{
run(script)
}
\arguments{
\item{script}{Name of the script (file in \verb{exec/})}
}
\description{
Used as Rscript -e 'phyfumr::run("scriptName.R")' --args arg1 arg2
}
