% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{simulate_betas}
\alias{simulate_betas}
\title{Simulates b-values}
\usage{
simulate_betas(counts, delta, eta, kappa)
}
\arguments{
\item{counts}{partitions::compositions(n,length(freqs))}

\item{delta}{flipflop.delta parameter of the phyfum model (methylated
background noise)}

\item{eta}{flipflop.eta parameter of the phyfum model (de-methylated
background negative noise)}

\item{kappa}{flipflop.kappa parameter of the phyfum model (b concentration)}
}
\value{
b-value alignment (i.e., matrix (or 2-dimensional array) with a
column per fCpG site and row per sample) like the one needed by
\code{\link[=get_peak_likelihoods]{get_peak_likelihoods()}}
}
\description{
Needed for the full-probability fss estimation
}
\details{
This is old code not currently used and has memory and speed issues unless S
is very due small since it pre-computes all the counts and does not discard
cases with very small multinomial probability
}
\seealso{
\code{\link[=get_fss_entropy]{get_fss_entropy()}}
}
\keyword{internal,}
\keyword{unsupported}
