% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{sumfun}
\alias{sumfun}
\title{Summary function for point estimates}
\usage{
sumfun(x, type = c("mean", "median", "mode"), ...)
}
\arguments{
\item{x}{data}

\item{type}{character indicating if we want to use the median, mean, or mode}

\item{...}{arguments for the corresponding summary function}
}
\value{
point estimate
}
\description{
Used for plotting
}
\seealso{
\code{\link[=plot_continuous_parameters]{plot_continuous_parameters()}}
}
