% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{tree_convergence_stats}
\alias{tree_convergence_stats}
\title{Generates a table with tree convergence statistics using the pseudo ESS
approach}
\usage{
tree_convergence_stats(
  chains,
  burnin_p = 0,
  n_focal_trees = 20,
  treedist = c("PD", "RF"),
  cred_mass = 0.95,
  by_chain = T,
  cl = 1
)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin. Calculated
by trace and flagged or removed}

\item{n_focal_trees}{Number of focal trees to calculate tree distances}

\item{treedist}{Tree distance to be used to calculate the distance to the
focal tree ("PD" or "RF")}

\item{cred_mass}{mass of the selected high posterior density interval (HDI)}

\item{by_chain}{Calculates Rhat, and the two ESS parameters both using all
chains and by chain (otherwise, using all of them only)}

\item{cl}{cl settings for \code{\link[pbapply:pbapply]{pbapply::pblapply()}} if available}
}
\value{
data.table with a row per parameter and colums for Rhat, ESS-bulk, and ESS-tail, median, mean, and HDI interval (HDILower and HDIUpper)
}
\description{
Generates a table with tree convergence statistics using the pseudo ESS
approach
}
\details{
The pseudo-ESS approach generates a trace by calculating a tree
distance to a focal tree (usually chosen at random). Then, that trace can
be used to assess mixing, and anything else typically done with continuous
parameters. To improve its accuracy, a number of replicates are generated
and mean convergence statistics reported.
}
\seealso{
\code{\link[=convergence_stats]{convergence_stats()}}
}
