% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{update_ptable}
\alias{update_ptable}
\title{Updates RWTY parameter tables for multiple MCMC chains}
\usage{
update_ptable(chains, thedata, exclude_cols = c("chain"))
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}

\item{thedata}{data.table with modified chain data}

\item{exclude_cols}{columns in thedata that should not be added to chains.
Defaults to those columns added by \code{\link[=extract_ptable]{extract_ptable()}} except burnin}
}
\value{
new chains
}
\description{
Data can be extracted using \code{\link[=extract_ptable]{extract_ptable()}}, modified, and then this
function will add it back to the chains
}
\keyword{internal}
