% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{user_priors}
\alias{user_priors}
\title{Returns the default Prior settings for user exploration and modification}
\usage{
user_priors()
}
\value{
list of lists with prior settings. The first level indicates the named prior configuration, the second the parameter name.
}
\description{
Returns the default Prior settings for user exploration and modification
}
\details{
The default settings have improper luca.branch priors so this function is needed to modify it if needed
}
