% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{write_continuous_parameter_plots}
\alias{write_continuous_parameter_plots}
\title{Generates a pdf file with plots for a list of continuous parameters}
\usage{
write_continuous_parameter_plots(
  thedata,
  params,
  out_dir,
  out_name,
  ndigits = 3,
  width = 7,
  height = 7
)
}
\arguments{
\item{thedata}{data.table, with burnin removed if needed}

\item{params}{array of parameter names}

\item{out_dir}{output directory}

\item{out_name}{output name}

\item{ndigits}{number of digits to round numbers printed in the plot}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Generates a pdf file with plots for a list of continuous parameters
}
