% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R, R/traces.R
\name{merge_traces}
\alias{merge_traces}
\title{Merges a list of RWTY chains into one}
\usage{
merge_traces(traces, burnin_p = 0)

merge_traces(traces, burnin_p = 0)
}
\arguments{
\item{burnin_p}{Proportion of MCMC samples to discard as burnin. Calculated
by trace and flagged or removed}
}
\value{
rwty.chain with the chains concatenated and the burnin states (trees
and parameters) removed.

list with trees and ptable with the traces concatenated and the burnin states (trees
and parameters) removed.
}
\description{
Merges a list of RWTY chains into one

Merges a list of PHYFUM traces into one
}
\details{
It assumes samples are taken at regular intervals to perform the
burnin

It assumes samples are taken at regular intervals to perform the
burnin
}
\keyword{internal}
