% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{model_average_traces}
\alias{model_average_traces}
\title{Perform model-averaging across PHYFUMr traces from different models}
\usage{
model_average_traces(
  traces,
  log_joint_probabilities = NULL,
  log_prior_probabilities = NULL,
  posterior_probabilities = NULL,
  burnin_p = 0,
  n_marginalized_posterior_samples = 10000,
  max_expected_samples = 1,
  warn_biased_marginalization = T
)
}
\arguments{
\item{traces}{MCMC data. List of \code{\link[=load_traces]{load_traces()}} values}

\item{log_joint_probabilities}{array of the log joint probabilities (log
marginal likelihoods + log prior) of each of the posterior samples (traces)}

\item{log_prior_probabilities}{array of log prior probabilities for each of
the models to average across (traces)}

\item{posterior_probabilities}{array of the posterior probabilities of each
(S) (i.e., resampling probabilities)}

\item{burnin_p}{Proportion of MCMC samples to discard as burnin}

\item{n_marginalized_posterior_samples}{length of the resampled posterior. If
NULL, it is maximized sampling without resampling. If more samples than
existing are required, sampling is done with replacement, otherwise
without.}

\item{max_expected_samples}{maximum allowed expected number of samples in an
extreme value of S}

\item{warn_biased_marginalization}{Logical, indicating whether to warn or not
if the limit of
max_expected_samples in an extreme value of S is violated}
}
\value{
model-averaged chain
}
\description{
Perform model-averaging across PHYFUMr traces from different models
}
\details{
If the joint probabilities are not provided, the marginal likelihood
is estimated here using the harmonic mean estimator. If the prior
probabilities for each model are provided, they are used to calculate the
final joint probabilities, otherwise, a uniform prior is implicitly used.
}
