% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_qc.R
\name{remove_burnin}
\alias{remove_burnin}
\title{Updates RWTY chains removing burnin states in parameter and tree objects for
multiple MCMC chains}
\usage{
remove_burnin(chains)
}
\arguments{
\item{chains}{MCMC data. List of \code{\link[rwty:load.trees]{rwty::load.trees()}} values}
}
\value{
new chains
}
\description{
Data can be extracted and burnin tagged using \code{\link[=extract_ptable]{extract_ptable()}}, modified
manually if desired, and added back using \code{\link[=update_ptable]{update_ptable()}}. This function,
removes burnin states marked in a column called burnin (T: removed, F: kept)
}
\keyword{internal}
