% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-utils.R
\name{slapply}
\alias{slapply}
\title{Uses pblapply or lapply depending on their availability}
\usage{
slapply(X, FUN, ...)
}
\arguments{
\item{X}{a vector (atomic or list) or an \code{\link[base]{expression}}
    object.  Other objects (including classed objects) will be coerced
    by \code{base::\link[base]{as.list}}.}

\item{FUN}{the function to be applied to each element of \code{X}:
    see \sQuote{Details}.  In the case of functions like
    \code{+}, \code{\%*\%}, the function name must be backquoted or quoted.}

\item{...}{
  Arguments passed on to \code{\link[pbapply:pbapply]{pbapply::pblapply}}
  \describe{
    \item{\code{cl}}{
A cluster object created by \code{\link{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations
(see Details on performance).
It can also be \code{"future"} to use a future backend (see Details),
\code{NULL} (default) refers to sequential evaluation.
}
  }}
}
\description{
Uses pblapply or lapply depending on their availability
}
\keyword{internal}
